% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-WGSRPD3.R
\docType{data}
\name{WGSRPD3}
\alias{WGSRPD3}
\title{Biodiversity Information Standards (TDWG) World Geographical Scheme for Recording Plant Distributions (WGSRPD)}
\format{
An 'sf' object with 20 rows and 5 variables:
\describe{
\item{LEVEL3_NAM}{Region name}
\item{LEVEL3_COD}{Region code}
\item{LEVEL2_COD}{Level 2 code}
\item{LEVEL1_COD}{Level 1 code (continent)}
\item{geometry}{sf geometry}
}
}
\source{
\itemize{
\item \href{https://github.com/tdwg/wgsrpd/tree/master/level3}{level3}
}
}
\usage{
WGSRPD3
}
\description{
Spatial data for WGSRPD Level 3
}
\references{
Govaerts, R., Nic Lughadha, E., Black, N. et al. The World Checklist of Vascular Plants, a continuously updated resource for exploring global plant diversity. \emph{Sci Data} 8, 215 (2021). \doi{10.1038/s41597-021-00997-6}
}
\keyword{datasets}
