% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Distributions.R
\docType{data}
\name{Distributions}
\alias{Distributions}
\title{The World Checklist of Vascular Plants Distributions}
\format{
A data.table with 1,983653 rows and 6 variables:
\describe{
\item{plant_name_id}{World Checklist of Vascular Plants (WCVP) identifier}
\item{taxon_name}{Concatenation of genus with species and, where applicable, infraspecific epithets to make a binomial or trinomial name}
\item{area_code_l3}{WGSRPD Level 3 code}
\item{introduced}{0 if native; 1 if introduced}
\item{extinct}{1 if extinct; 0 if extant}
\item{location_doubtful}{1 if doubtful; 0 otherwise}
}
}
\source{
\url{https://sftp.kew.org/pub/data-repositories/WCVP/wcvp.zip}
}
\usage{
Distributions
}
\description{
A processed dataset containing the distribution data from the WCVP, mapped to the
Biodiversity Information Standards (TDWG) World Geographical Scheme for
Recording Plant Distributions (WGSRPD)
}
\references{
Govaerts, R., Nic Lughadha, E., Black, N. et al. The World Checklist of Vascular Plants, a continuously updated resource for exploring global plant diversity. \emph{Sci Data} 8, 215 (2021). \doi{10.1038/s41597-021-00997-6}
}
\keyword{datasets}
