% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-WorldLandMap.R
\docType{data}
\name{WorldLandMap}
\alias{WorldLandMap}
\title{A simple features object of the world land map}
\format{
A simple features object
\describe{
\item{featurecla}{Land area}
\item{scalerank}{Scale rank of the land feature}
\item{min_zoom}{Minimum zoom level at which the feature is displayed}
\item{geometry}{Geometry unit}
}
}
\source{
\code{rnaturalearth::ne_download(scale = 110,type = 'land',category = 'physical',returnclass = "sf")}
}
\usage{
WorldLandMap
}
\description{
An integrated land map
}
\seealso{
\code{\link[rnaturalearth:ne_download]{rnaturalearth::ne_download()}}
}
\keyword{datasets}
